AaltoFloatsEdge/modules/dvlmodule/requirements.txt
azure-iot-device~=2.7.0
flatten-json~=0.1.13